Adding sounddevices testfile for cue points
As described in issue #9 the function cue_points() didn't display the
expected frame values for cue points created on a Sound Devices
MixPre-Series recorder. The file added by this commit
(sounddevices_6_cue_points.wav) is short and mono channel, but has 6 cue
points in it.

The expected frames for the cue points are:
0: 90112
1: 176128
2: 237568
3: 294912
4: 380928
5: 385024